requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
aiosqlite>=0.20.0
google-genai
python-dotenv>=1.0.1
jinja2>=3.1.2
chromadb>=0.4.22
